Backend part Done

+++ backend/controllers/listing.js
@@ -1,4 +1,6 @@
 const { Listing, validate } = require("../models/Listing");
+const CustomError = require('../utils/customError')
+
 
 //Get All Listing Houses
 const getallListing = async (req, res) => {
@@ -6,19 +8,20 @@ const getallListing = async (req, res) => {
   let queary;
   
   if(req.query){
-    const {address,city} = req.query
+    const {address,city,sort} = req.query
     const quaryObject = {}
   
     if(address){
       quaryObject.address = address
     }
     if(city){
-      quaryObject.city = city
+      quaryObject.city = { $regex : name, $options: 'i'}
     }
-  
-     queary = Listing.find(quaryObject)
+     
+    queary = Listing.find(quaryObject)
+
   }else{
-    queary = Listing.find({})
+    queary = Listing.find()
   }
  
 
@@ -32,7 +35,7 @@ const getallListing = async (req, res) => {
   });
 };
 
-//Get Getail House Listing
+//Get Get One House Listing
 const getListing = async (req, res) => {
 
   let id = req.params.id;
@@ -50,6 +53,8 @@ const postlListing = async (req, res) => {
 
   let { title, address, city, description, price, rooms, bedroom } = req.body;
 
+  req.body.publisher = req.user._id;
+
   let { error, value } = validate(
     title,
     address,
@@ -61,10 +66,7 @@ const postlListing = async (req, res) => {
   );
 
   if (error) {
-    return res.status(400).json({
-      success: false,
-      error: error.message,
-    });
+    throw new CustomError(error.message,400)
   }
 
   const listing = await Listing.create(req.body);
@@ -83,11 +85,15 @@ const updateListing = async (req, res) => {
   let listing = await Listing.findById(id);
 
   if (!listing) {
-    return res.status(400).json({
-      success: false,
-      error: "No Listing with given id",
-    });
+    throw new CustomError("No Listing With Given Id",404)
   }
+
+
+  //Check whether login User Is Publisher of the Listing
+  if(listing.publisher.toString() !== req.user.id){
+   throw new CustomError("Only Author can Update Listing",401)
+  }
+ 
 
   let { title, address, city, description, price, rooms, bedroom } = req.body;
 
@@ -102,10 +108,7 @@ const updateListing = async (req, res) => {
   );
 
   if (error) {
-    return res.status(400).json({
-      success: false,
-      error: error.message,
-    });
+    throw new CustomError(error.message,404)
   }
 
   listing = await Listing.findByIdAndUpdate(id, req.body, {
@@ -127,11 +130,14 @@ const removelListing = async (req, res) => {
     let listing = await Listing.findById(id)
 
     if(!listing){
-       return res.status(400).json({
-            success:false,
-            error:'No Listing with Given Id'
-        })
+      throw new CustomError("No Listing With Given Id",404)
     }
+
+     //Check whether login User Is Publisher of the Listing
+    if(listing.publisher.toString() !== req.user.id){
+      throw new CustomError("Only Author can Update Listing",401)
+    }
+   
 
     await Listing.findByIdAndRemove(id)
 
@@ -144,6 +150,22 @@ const removelListing = async (req, res) => {
 };
 
 
+const uplaodImage = async (req,res,next) => {
+  if(!req.file){
+    throw new CustomError("Please upload Photo",400)
+  }
+  if(!req.file.mimetype.startsWith("image")){
+    throw new CustomError("File format not valid",400);
+  }
+  req.file.filename = req.user.id
+  
+  res.status(200).json({
+    success:true,
+    data: req.file.filename
+  })
+  
+
+}
 
 
 
@@ -155,5 +177,6 @@ module.exports = {
   postlListing,
   updateListing,
   removelListing,
+  uplaodImage
 };
